replace DBDebugKit with simple ViewController showing our custom debugging variables
